fix: #114 incorrect span for operators.
This patch fixes improper generated spans for operators and
in general cleans up the sources surrounding operators.
Everything operator related (frontend agnostic) was moved into
`hash-ast` and replaced the old operator types.
