+++ compiler/hash-parser/src/gen.rs
@@ -3,32 +3,26 @@
 //!
 //! All rights reserved 2021 (c) The Hash Language authors
 
-use std::cell::Cell;
-use std::path::PathBuf;
-use std::str::FromStr;
-
 use hash_alloc::collections::row::Row;
 use hash_alloc::{row, Wall};
-use hash_ast::ast_nodes;
-use hash_ast::ident::CORE_IDENTIFIERS;
 use hash_ast::{
     ast::*,
     ident::{Identifier, IDENTIFIER_MAP},
     keyword::Keyword,
     literal::STRING_LITERAL_MAP,
-    operator::{CompoundFn, OperatorFn},
 };
+use hash_ast::{ast_nodes, operator::Operator};
+use hash_ast::{ident::CORE_IDENTIFIERS, operator::OperatorKind};
 use hash_source::location::{Location, SourceLocation};
+use std::cell::Cell;
+use std::path::PathBuf;
+use std::str::FromStr;
 
 use crate::parser::ImportResolver;
+use crate::token::{Delimiter, Token, TokenKindVector};
 use crate::{
     error::{AstGenError, AstGenErrorKind, TyArgumentKind},
-    operator::Operator,
     token::TokenKind,
-};
-use crate::{
-    operator::OperatorKind,
-    token::{Delimiter, Token, TokenKindVector},
 };
 
 pub type AstGenResult<'a, T> = Result<T, AstGenError<'a>>;
@@ -535,16 +529,11 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
         lhs: AstNode<'c, Expression<'c>>,
     ) -> AstGenResult<'c, (AstNode<'c, Expression<'c>>, bool)> {
         if let Some(op) = self.peek_resultant_fn(|| self.parse_re_assignment_op()) {
-            let transformed_op: OperatorFn = op.into();
-
             // Parse the rhs and the semi
             let rhs = self.parse_expression_with_precedence(0)?;
 
             // Now we need to transform the re-assignment operator into a function call
-            Ok((
-                self.transform_binary_expression(lhs, rhs, transformed_op),
-                false,
-            ))
+            Ok((self.transform_binary_expression(lhs, rhs, op), false))
         } else {
             Ok((lhs, false))
         }
@@ -592,55 +581,43 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
         &self,
         lhs: AstNode<'c, Expression<'c>>,
         rhs: AstNode<'c, Expression<'c>>,
-        op: OperatorFn,
+        op: AstNode<'c, Operator>,
     ) -> AstNode<'c, Expression<'c>> {
-        match op {
-            OperatorFn::Named { name, assigning } => {
-                let rhs_location = rhs.location();
+        let Operator { kind, assignable } = op.body();
 
-                self.node(Expression::new(ExpressionKind::FunctionCall(
+        if kind.is_compound() {
+            // for compound functions that include ordering, we essentially transpile
+            // into a match block that checks the result of the 'ord' fn call to the
+            // 'Ord' enum variants. This also happens for operators such as '>=' which
+            // essentially means that we have to check if the result of 'ord()' is either
+            // 'Eq' or 'Gt'.
+            self.transform_compound_ord_fn(*kind, *assignable, lhs, rhs)
+        } else if kind.is_lazy() {
+            // some functions have to exhibit a short-circuiting behaviour, namely
+            // the logical 'and' and 'or' operators. To do this, we expect the 'and'
+            // 'or' trait (and their assignment counterparts) to expect the rhs part
+            // as a lambda. So, we essentially create a lambda that calls the rhs, or
+            // in other words, something like this happens:
+            //
+            // >>> lhs && rhs
+            // vvv (transpiles to...)
+            // >>> and(lhs, () => rhs)
+            //
+            let location = lhs.location().join(rhs.location());
+
+            self.node_from_location(Expression::new(ExpressionKind::FunctionCall(
                     FunctionCallExpr {
-                        subject: self.node(Expression::new(ExpressionKind::Variable(
-                            VariableExpr {
-                                name: self.make_access_name_from_str(name, self.current_location()),
+                        subject: self.node_from_location(
+                            Expression::new(ExpressionKind::Variable(VariableExpr {
+                                name: self.make_access_name_from_str(kind.to_string(), op.location()),
                                 type_args: AstNodes::empty(),
-                            },
-                        ))),
+                            })),
+                            &op.location(),
+                        ),
                         args: self.node_from_joined_location(
                             FunctionCallArgs {
-                                entries: ast_nodes![&self.wall;
-                                    self.transform_expr_into_ref(lhs, assigning),
-                                    rhs,
-                                ],
-                            },
-                            &rhs_location,
-                        ),
-                    },
-                )))
-            }
-            OperatorFn::LazyNamed { name, assigning } => {
-                // some functions have to exhibit a short-circuiting behaviour, namely
-                // the logical 'and' and 'or' operators. To do this, we expect the 'and'
-                // 'or' trait (and their assignment counterparts) to expect the rhs part
-                // as a lambda. So, we essentially create a lambda that calls the rhs, or
-                // in other words, something like this happens:
-                //
-                // >>> lhs && rhs
-                // vvv (transpiles to...)
-                // >>> and(lhs, () => rhs)
-                //
-
-                self.node(Expression::new(ExpressionKind::FunctionCall(
-                    FunctionCallExpr {
-                        subject: self.node(Expression::new(ExpressionKind::Variable(
-                            VariableExpr {
-                                name: self.make_access_name_from_str(name, self.current_location()),
-                                type_args: AstNodes::empty(),
-                            },
-                        ))),
-                        args: self.node(FunctionCallArgs {
                             entries: ast_nodes![&self.wall;
-                                self.transform_expr_into_ref(lhs, assigning),
+                                self.transform_expr_into_ref(lhs, *assignable),
                                 self.node(Expression::new(ExpressionKind::LiteralExpr(LiteralExpr(self.node(
                                     Literal::Function(FunctionDef {
                                         args: AstNodes::empty(),
@@ -649,24 +626,39 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
                                     }),
                                 )))))
                             ],
-                        }),
+                        },  &location),
                     },
-                )))
-            }
-            OperatorFn::Compound { name, assigning } => {
-                // for compound functions that include ordering, we essentially transpile
-                // into a match block that checks the result of the 'ord' fn call to the
-                // 'Ord' enum variants. This also happens for operators such as '>=' which
-                // essentially means that we have to check if the result of 'ord()' is either
-                // 'Eq' or 'Gt'.
-                self.transform_compound_ord_fn(name, assigning, lhs, rhs)
-            }
+                )), &location)
+        } else {
+            let location = lhs.location().join(rhs.location());
+
+            self.node_from_location(
+                Expression::new(ExpressionKind::FunctionCall(FunctionCallExpr {
+                    subject: self.node_from_location(
+                        Expression::new(ExpressionKind::Variable(VariableExpr {
+                            name: self.make_access_name_from_str(kind.to_string(), op.location()),
+                            type_args: AstNodes::empty(),
+                        })),
+                        &op.location(),
+                    ),
+                    args: self.node_from_joined_location(
+                        FunctionCallArgs {
+                            entries: ast_nodes![&self.wall;
+                                self.transform_expr_into_ref(lhs, *assignable),
+                                rhs,
+                            ],
+                        },
+                        &location,
+                    ),
+                })),
+                &op.location(),
+            )
         }
     }
 
     fn transform_compound_ord_fn(
         &self,
-        fn_ty: CompoundFn,
+        fn_ty: OperatorKind,
         assigning: bool,
         lhs: AstNode<'c, Expression<'c>>,
         rhs: AstNode<'c, Expression<'c>>,
@@ -689,7 +681,7 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
         // should return 'true' on, and all the rest will return false...
         // the order is (Lt, Eq, Gt)
         let mut branches = match fn_ty {
-            CompoundFn::Leq => {
+            OperatorKind::LtEq => {
                 ast_nodes![&self.wall; self.node(MatchCase {
                     pattern: self.node(Pattern::Or(OrPattern {
                         variants: ast_nodes![&self.wall;
@@ -700,7 +692,7 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
                     expr: self.make_variable(self.make_boolean(true)),
                 })]
             }
-            CompoundFn::Geq => {
+            OperatorKind::GtEq => {
                 ast_nodes![&self.wall; self.node(MatchCase {
                     pattern: self.node(Pattern::Or(OrPattern {
                         variants: ast_nodes![&self.wall;
@@ -711,18 +703,19 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
                     expr: self.make_variable(self.make_boolean(true)),
                 })]
             }
-            CompoundFn::Lt => {
+            OperatorKind::Lt => {
                 ast_nodes![&self.wall; self.node(MatchCase {
                     pattern: self.make_enum_pattern_from_str("Lt", location),
                     expr: self.make_variable(self.make_boolean(false)),
                 })]
             }
-            CompoundFn::Gt => {
+            OperatorKind::Gt => {
                 ast_nodes![&self.wall; self.node(MatchCase {
                     pattern: self.make_enum_pattern_from_str("Gt", location),
                     expr: self.make_variable(self.make_boolean(false)),
                 })]
             }
+            _ => unreachable!(),
         };
 
         // add the '_' case to the branches to return false on any other
@@ -778,8 +771,9 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
     /// This will parse an operator and check that it is re-assignable, if the operator is not
     /// re-assignable then the result of the function is an [AstGenError] since it expects there
     /// to be this operator.
-    pub fn parse_re_assignment_op(&self) -> AstGenResult<'c, Operator> {
-        let (operator, consumed_tokens) = Operator::from_token_stream(self);
+    pub fn parse_re_assignment_op(&self) -> AstGenResult<'c, AstNode<'c, Operator>> {
+        let start = self.next_location();
+        let (operator, consumed_tokens) = self.parse_operator();
 
         match operator {
             Some(Operator {
@@ -789,10 +783,13 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
                 // consume the number of tokens eaten whilst getting the operator...
                 self.offset.update(|x| x + consumed_tokens as usize);
 
-                Ok(Operator {
-                    kind,
-                    assignable: true,
-                })
+                Ok(self.node_from_joined_location(
+                    Operator {
+                        kind,
+                        assignable: true,
+                    },
+                    &start,
+                ))
             }
             _ => self.error(AstGenErrorKind::ReAssignmentOp, None, None), // TODO: actually add information here
         }
@@ -1789,9 +1786,6 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
             // check for assigning operators here if the lhs expression is not compound
             match gen.peek_resultant_fn(|| gen.parse_re_assignment_op()) {
                 Some(op) => {
-                    // transform the operator kind into OperatorFn so we can transform the operator into a function call.
-                    let transformed_op: OperatorFn = op.into();
-
                     // since this is followed by an expression, we try to parse another expression, and then
                     // ensure that after an expression there is a ending semi colon.
                     let rhs = gen.parse_expression_with_precedence(0)?;
@@ -1799,11 +1793,9 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
 
                     block.statements.nodes.push(
                         gen.node_from_joined_location(
-                            Statement::Expr(ExprStatement(self.transform_binary_expression(
-                                expr,
-                                rhs,
-                                transformed_op,
-                            ))),
+                            Statement::Expr(ExprStatement(
+                                self.transform_binary_expression(expr, rhs, op),
+                            )),
                             &expr_loc,
                         ),
                         &self.wall,
@@ -2633,10 +2625,10 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
         self.is_compound_expr.set(false);
 
         loop {
-            let op_start = self.current_location();
+            let op_start = self.next_location();
             // this doesn't consider operators that have an 'eq' variant because that is handled at the statement level,
             // since it isn't really a binary operator...
-            let (op, consumed_tokens) = Operator::from_token_stream(self);
+            let (op, consumed_tokens) = self.parse_operator();
 
             match op {
                 // check if the operator here is re-assignable, as in '+=', '/=', if so then we need to stop
@@ -2675,8 +2667,11 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
                         self.is_compound_expr.set(true);
 
                         // transform the operator into an OperatorFn
-                        let transformed_op: OperatorFn = op.into();
-                        lhs = self.transform_binary_expression(lhs, rhs, transformed_op);
+                        lhs = self.transform_binary_expression(
+                            lhs,
+                            rhs,
+                            self.node_with_location(op, op_span),
+                        );
                     }
                 }
                 _ => break,
@@ -3499,5 +3494,78 @@ impl<'c, 'stream, 'resolver> AstGen<'c, 'stream, 'resolver> {
             Expression::new(ExpressionKind::LiteralExpr(LiteralExpr(literal))),
             &token.span,
         )
+    }
+
+    /// This function is used to pickup 'glued' operator tokens to form more complex binary operators
+    /// that might be made up of multiple tokens. The function will peek ahead (2 tokens at most since
+    /// all binary operators are made of that many tokens). The function returns an optional derived
+    /// operator, and the number of tokens that was consumed deriving the operator, it is the responsibility
+    /// of the caller to increment the token stream by the provided number.
+    pub(crate) fn parse_operator(&self) -> (Option<Operator>, u8) {
+        let token = self.peek();
+
+        // check if there is a token that we can peek at ahead...
+        if token.is_none() {
+            return (None, 0);
+        }
+
+        let (op, mut consumed): (_, u8) = match &(token.unwrap()).kind {
+            // Since the 'as' keyword is also a binary operator, we have to handle it here...
+            TokenKind::Keyword(Keyword::As) => (Some(OperatorKind::As), 1),
+            TokenKind::Eq => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Eq => (Some(OperatorKind::EqEq), 2),
+                _ => (None, 0),
+            },
+            TokenKind::Lt => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Eq => (Some(OperatorKind::LtEq), 2),
+                Some(token) if token.kind == TokenKind::Lt => (Some(OperatorKind::Shl), 2),
+                _ => (Some(OperatorKind::Lt), 1),
+            },
+            TokenKind::Gt => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Eq => (Some(OperatorKind::GtEq), 2),
+                Some(token) if token.kind == TokenKind::Gt => (Some(OperatorKind::Shr), 2),
+                _ => (Some(OperatorKind::Gt), 1),
+            },
+            TokenKind::Plus => (Some(OperatorKind::Add), 1),
+            TokenKind::Minus => (Some(OperatorKind::Sub), 1),
+            TokenKind::Star => (Some(OperatorKind::Mul), 1),
+            TokenKind::Slash => (Some(OperatorKind::Div), 1),
+            TokenKind::Percent => (Some(OperatorKind::Mod), 1),
+            TokenKind::Caret => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Caret => (Some(OperatorKind::Exp), 2),
+                _ => (Some(OperatorKind::BitXor), 1),
+            },
+            TokenKind::Amp => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Amp => (Some(OperatorKind::And), 2),
+                _ => (Some(OperatorKind::BitAnd), 1),
+            },
+            TokenKind::Pipe => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Pipe => (Some(OperatorKind::Or), 2),
+                _ => (Some(OperatorKind::BitOr), 1),
+            },
+            TokenKind::Exclamation => match self.peek_second() {
+                Some(token) if token.kind == TokenKind::Eq => (Some(OperatorKind::NotEq), 2),
+                _ => (None, 0), // this is a unary operator '!'
+            },
+            _ => (None, 0),
+        };
+
+        match op {
+            // check if the operator is re-assignable and if so try to parse a equality operator.
+            // If we see this one then we can parse and then just convert the operator into the
+            // 'Eq' version.
+            Some(kind) => {
+                let assignable = match self.peek_nth(consumed as usize) {
+                    Some(token) if kind.is_re_assignable() && token.has_kind(TokenKind::Eq) => {
+                        consumed += 1;
+                        true
+                    }
+                    _ => false,
+                };
+
+                (Some(Operator { kind, assignable }), consumed)
+            }
+            None => (None, 0),
+        }
     }
 }
